Can now add items to the canvas from JS calls
This is a pretty shitty hack. Instead of passing messages directly, I'm
pushing the items to an attribute of a hidden div. The app then checks
that attribute every time it updates to see if three are new items, and
then sends the add item message to itself for each item. The inital page
update is called from js by clicking a button that is inside that hidden
div